fix some warnings

+++ plugins/b2slegacy/collections/ControlCollection.cpp
@@ -30,8 +30,8 @@ void ControlCollection::Add(ControlInfo* pControlInfo)
 void ControlCollection::SetScore(B2SData* pB2SData, int score)
 {
    string scoreAsString = string(m_digits > std::to_string(score).length()
-      ? m_digits - std::to_string(score).length() : 0,  ' ') + std::to_string(score);
-   if (scoreAsString.length() > m_digits)
+      ? m_digits - (int)std::to_string(score).length() : 0,  ' ') + std::to_string(score);
+   if ((int)scoreAsString.length() > m_digits)
       scoreAsString = scoreAsString.substr(scoreAsString.length() - m_digits);
 
    for (auto& pControl : *this) {
